Added test for handling multi_search errors
Multi-search doesn't *always* return errors in the standard
format, it returns an array of responses, rather than a single
response. For this reason, we need to collect and flatten all
of the error arrays into one.

In addition to adding this spec, I had to add the concept of a
"static" engine for operations that require an indexed document. It's
"static" because it happens one time at the start of the test suite
and remains available for any read-only operation.

This is required because it takes time for a document to be indexed, so
we have to poll the API until the indexing is finished in order to
continue. It's better to do this once so that we only wait once,
not multiple times throughout the test suite.